+++ reiser4lib/libaal/math.c
@@ -8,10 +8,10 @@ int aal_pow_of_two(unsigned long n) {
     return (n & -n) == n;
 }
 
-int aal_log2(int n) {
-    int x;
+int aal_log2(unsigned long n) {
+    unsigned long x;
 
-    for (x = 0; 1 << x <= n; x++);
+    for (x = 0; (unsigned long)(1 << x) <= n; x++);
 	return x - 1;
 }
